requirements.txt
aes128==0.0.3
bcrypt==4.1.1
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==41.0.5
encode==0.0.0
flake8==7.0.0
flake8-html==0.4.3
greenlet==3.0.1
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
jaraco.classes==3.3.0
Jinja2==3.1.3
keyring==24.2.0
MarkupSafe==2.1.5
maskpass==0.3.7
mccabe==0.7.0
more-itertools==10.1.0
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pynput==1.7.6
pyOpenSSL==23.3.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
sentry-sdk==1.40.6
six==1.16.0
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
urllib3==2.1.0
wheel==0.41.2
pytest==8.1.1
